# SpringBoot-Rest-App-AirlineTicket/swagger-specification.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"Issue Management API Reference"},"host":"localhost:8080","basePath":"/","tags":[{"name":"ticket-controller","description":"Ticket Controller"},{"name":"company-controller","description":"Company Controller"},{"name":"flight-controller","description":"Flight Controller"},{"name":"plane-controller","description":"Plane Controller"},{"name":"route-controller","description":"Route Controller"},{"name":"airport-controller","description":"Airport Controller"}],"paths":{"/api/airport":{"get":{"tags":["airport-controller"],"summary":"Get All Operation","operationId":"getAllUsingGET","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Airport Entity Object"}}}}},"post":{"tags":["airport-controller"],"summary":"Create Operation","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"airport","description":"airport","required":true,"schema":{"$ref":"#/definitions/Airport Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Airport Entity Object"}}}},"put":{"tags":["airport-controller"],"summary":"Update Operation","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"airport","description":"airport","required":true,"schema":{"$ref":"#/definitions/Airport Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Airport Entity Object"}}}}},"/api/airport/{id}":{"get":{"tags":["airport-controller"],"summary":"Get By Id Operation","operationId":"getByIdUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Airport Entity Object"}}}},"delete":{"tags":["airport-controller"],"summary":"Delete Operation","operationId":"deleteUsingDELETE","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/company":{"get":{"tags":["company-controller"],"summary":"Get All Operation","operationId":"getAllUsingGET_1","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Company Entity Object"}}}}},"post":{"tags":["company-controller"],"summary":"Create Operation","operationId":"createUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"company","description":"company","required":true,"schema":{"$ref":"#/definitions/Company Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Company Entity Object"}}}}},"/api/company/{id}":{"get":{"tags":["company-controller"],"summary":"Get By Id Operation","operationId":"getByIdUsingGET_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Company Entity Object"}}}},"put":{"tags":["company-controller"],"summary":"Update Operation","operationId":"updateUsingPUT_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"company","description":"company","required":true,"schema":{"$ref":"#/definitions/Company Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Company Entity Object"}}}},"delete":{"tags":["company-controller"],"summary":"Delete Operation","operationId":"deleteUsingDELETE_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"boolean"}}}}},"/api/flight":{"get":{"tags":["flight-controller"],"summary":"Get All Operation","operationId":"getAllUsingGET_2","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Flight Entity Object"}}}}},"post":{"tags":["flight-controller"],"summary":"Create Operation","operationId":"createUsingPOST_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"flight","description":"flight","required":true,"schema":{"$ref":"#/definitions/Flight Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Flight Entity Object"}}}}},"/api/flight/{id}":{"get":{"tags":["flight-controller"],"summary":"Get By Id Operation","operationId":"getByIdUsingGET_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Flight Entity Object"}}}},"put":{"tags":["flight-controller"],"summary":"Update Operation","operationId":"updateUsingPUT_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"flight","description":"flight","required":true,"schema":{"$ref":"#/definitions/Flight Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Flight Entity Object"}}}},"delete":{"tags":["flight-controller"],"summary":"Delete Operation","operationId":"deleteUsingDELETE_2","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/plane":{"get":{"tags":["plane-controller"],"summary":"Get All Operation","operationId":"getAllUsingGET_3","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Plane Entity Object"}}}}},"post":{"tags":["plane-controller"],"summary":"Create Operation","operationId":"createUsingPOST_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"plane","description":"plane","required":true,"schema":{"$ref":"#/definitions/Plane Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Plane Entity Object"}}}}},"/api/plane/{id}":{"get":{"tags":["plane-controller"],"summary":"Get By Id Operation","operationId":"getByIdUsingGET_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Plane Entity Object"}}}},"put":{"tags":["plane-controller"],"summary":"Update Operation","operationId":"updateUsingPUT_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"plane","description":"plane","required":true,"schema":{"$ref":"#/definitions/Plane Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Plane Entity Object"}}}},"delete":{"tags":["plane-controller"],"summary":"Delete Operation","operationId":"deleteUsingDELETE_3","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/route":{"get":{"tags":["route-controller"],"summary":"Get All Operation","operationId":"getAllUsingGET_4","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Route Entity Object"}}}}},"post":{"tags":["route-controller"],"summary":"Create Operation","operationId":"createUsingPOST_4","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"route","description":"route","required":true,"schema":{"$ref":"#/definitions/Route Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Route Entity Object"}}}}},"/api/route/{id}":{"get":{"tags":["route-controller"],"summary":"Get By Id Operation","operationId":"getByIdUsingGET_4","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Route Entity Object"}}}},"put":{"tags":["route-controller"],"summary":"Update Operation","operationId":"updateUsingPUT_4","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"route","description":"route","required":true,"schema":{"$ref":"#/definitions/Route Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Route Entity Object"}}}},"delete":{"tags":["route-controller"],"summary":"Delete Operation","operationId":"deleteUsingDELETE_4","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}},"/api/ticket":{"get":{"tags":["ticket-controller"],"summary":"Get All Operation","operationId":"getAllUsingGET_5","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Ticket Entity Object"}}}}},"post":{"tags":["ticket-controller"],"summary":"Create Operation","operationId":"createUsingPOST_5","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"ticket","description":"ticket","required":true,"schema":{"$ref":"#/definitions/Ticket Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Ticket Entity Object"}}}}},"/api/ticket/{id}":{"get":{"tags":["ticket-controller"],"summary":"Get By Id Operation","operationId":"getByIdUsingGET_5","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Ticket Entity Object"}}}},"put":{"tags":["ticket-controller"],"summary":"Update Operation","operationId":"updateUsingPUT_5","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"ticket","description":"ticket","required":true,"schema":{"$ref":"#/definitions/Ticket Entity Object"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Ticket Entity Object"}}}},"delete":{"tags":["ticket-controller"],"summary":"Delete Operation","operationId":"deleteUsingDELETE_5","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object"}}}}}},"definitions":{"Ticket Entity Object":{"type":"object","required":["ticketCode"],"properties":{"flight":{"$ref":"#/definitions/Flight Entity Object"},"id":{"type":"integer","format":"int64","description":"Ticket Id"},"isSold":{"type":"boolean","example":false,"description":"Is Ticket Sold"},"price":{"type":"integer","format":"int32","description":"Price of Ticket"},"ticketCode":{"type":"string","description":"Code of Ticket"}}},"Company Entity Object":{"type":"object","required":["name"],"properties":{"flights":{"type":"array","items":{"$ref":"#/definitions/Flight Entity Object"}},"id":{"type":"integer","format":"int64","description":"Company Id"},"name":{"type":"string","description":"Name of Company"},"planes":{"type":"array","items":{"$ref":"#/definitions/Plane Entity Object"}}}},"Flight Entity Object":{"type":"object","required":["name"],"properties":{"arrivalDateTime":{"type":"string","format":"date-time","description":"Arrival Date and Time"},"company":{"$ref":"#/definitions/Company Entity Object"},"departureDateTime":{"type":"string","format":"date-time","description":"Departure Date and Time"},"duration":{"type":"integer","format":"int32","description":"Flight Duration"},"id":{"type":"integer","format":"int64","description":"Flight Id"},"name":{"type":"string","description":"Name of Flight"},"quota":{"type":"integer","format":"int32","description":"Flight Quota"},"quotaFilled":{"type":"integer","format":"int32","description":"Flight Quota Filled"},"quotaFilledPercentage":{"type":"integer","format":"int32","description":"Flight Quota Percentage"},"route":{"$ref":"#/definitions/Route Entity Object"},"tickets":{"type":"array","items":{"$ref":"#/definitions/Ticket Entity Object"}}}},"Route Entity Object":{"type":"object","required":["name"],"properties":{"arrivalAirport":{"$ref":"#/definitions/Airport Entity Object"},"departureAirport":{"$ref":"#/definitions/Airport Entity Object"},"flights":{"type":"array","items":{"$ref":"#/definitions/Flight Entity Object"}},"id":{"type":"integer","format":"int64","description":"Route Id"},"name":{"type":"string","description":"Name of Route"}}},"Airport Entity Object":{"type":"object","required":["name"],"properties":{"arrivalRoutes":{"type":"array","items":{"$ref":"#/definitions/Route Entity Object"}},"capacity":{"type":"integer","format":"int32","description":"Capacity of Airport"},"departureRoutes":{"type":"array","items":{"$ref":"#/definitions/Route Entity Object"}},"id":{"type":"integer","format":"int64","description":"Airport Id"},"location":{"type":"string","description":"Location of Airport"},"name":{"type":"string","description":"Name of Airport"}}},"Plane Entity Object":{"type":"object","required":["name"],"properties":{"company":{"$ref":"#/definitions/Company Entity Object"},"id":{"type":"integer","format":"int64","description":"Plane Id"},"name":{"type":"string","description":"Name of Plane"},"numberOfSeats":{"type":"integer","format":"int32","description":"Number Of Seats"}}}}}